Added initial csv work
